renamed utils to texture.h
